DX12: improved the profiler readback, change RG interface, scrolling in ImGui

+++ DX12/D3D12Research/D3D12/Graphics/Core/GraphicsBuffer.h
@@ -44,7 +44,7 @@ struct BufferDesc
 
 	static BufferDesc CreateReadback(uint32_t size)
 	{
-		return BufferDesc(size, sizeof(uint64_t), BufferFlag::Readback);
+		return BufferDesc(size, sizeof(uint8_t), BufferFlag::Readback);
 	}
 
 	static BufferDesc CreateStructured(uint32_t elementCount, uint32_t elementSize, BufferFlag usage = BufferFlag::ShaderResource | BufferFlag::UnorderedAccess)
